Move the definition of the custom UnaccentCharField in a general models.py in openquakeplatform
