+++ openquakeplatform/openquakeplatform/vulnerability/models.py
@@ -26,50 +26,7 @@ from commons import (SemicolonSeparatedFloatFormField,
 #from django.core import validators
 #from django.core.validators import RegexValidator
 from django.core.validators import MinValueValidator, MaxValueValidator
-
-from django.db.models.sql.constants import QUERY_TERMS
-from django.db.backends.postgresql_psycopg2.base import DatabaseWrapper
-from django.db.backends.postgresql_psycopg2.operations import (
-    DatabaseOperations)
-
-QUERY_TERMS.add('unaccent')
-DatabaseWrapper.operators['unaccent'] = '=~@ %s'
-
-
-def my_lookup_cast(self, lookup_type):
-    # Adding 'unaccent' to the standard lookup_cast
-    lookup = '%s'
-
-    # Cast text lookups to text to allow things like filter(x__contains=4)
-    if lookup_type in ('iexact', 'contains', 'icontains', 'unaccent',
-                       'startswith', 'istartswith', 'endswith', 'iendswith'):
-        lookup = "%s::text"
-
-    # Use UPPER(x) for case-insensitive lookups; it's faster.
-    if lookup_type in ('iexact', 'icontains', 'unaccent',
-                       'istartswith', 'iendswith'):
-        lookup = 'UPPER(%s)' % lookup
-
-    return lookup
-
-DatabaseOperations.lookup_cast = my_lookup_cast
-
-
-class UnaccentCharField(models.CharField):
-    def get_db_prep_lookup(self, lookup_type, value,
-                           connection, prepared=False):
-        if lookup_type == 'unaccent':
-            return ["%%%s%%" % connection.ops.prep_for_like_query(value)]
-        else:
-            return super(UnaccentCharField, self).get_db_prep_lookup(
-                lookup_type, value, connection, prepared)
-
-    def get_prep_lookup(self, lookup_type, value):
-        if lookup_type == 'unaccent':
-            return value
-        else:
-            return super(UnaccentCharField, self).get_prep_lookup(lookup_type,
-                                                                  value)
+from openquakeplatform.models import UnaccentCharField
 
 
 class SemicolonSeparatedFloatField(models.CharField):
